refactor(sdk): Rename `Persistence.ts` to `types.ts` in persistence utils
- Move Persistence interface from Persistence.ts to types.ts
- Update import in PersistenceManager to use type-only import
